pyproject.toml
[tool.poetry]
name = "pd15-ishemgulov-lesson28"
version = "0.1.0"
description = ""
authors = ["rizv4n <rizvanishemgulov@gmail.com>"]
readme = "README.md"
packages = [{include = "pd15_ishemgulov_lesson28"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
